:sparkles: install script

+++ install.template.sh
@@ -8,15 +8,21 @@ set -eu
 GH_USER="{{GH_USER}}"
 PROJECT="{{PROJECT}}"
 ENTRYPOINT="{{ENTRYPOINT}}"
+
 BIN_DIR="$HOME/.local/bin"
 ENVS_DIR="$HOME/.local/envs"
 
+# Either empty or set from user input:
+VERSION=""
+NAME=""
 
 usage() {
   echo "Usage: $0 [options]"
   echo "  [--gh-user GH_USER]         GitHub user (default: $GH_USER)"
   echo "  [--project PROJECT]         Project name (default: $PROJECT)"
   echo "  [--entrypoint ENTRYPOINT]   Executable name to symlink as (default: $ENTRYPOINT)"
+  echo "  [--name NAME]               Name of the symlink (default: $NAME)"
+  echo "  [--version VERSION]         Version to install (default: latest from GitHub)"
   echo "  [--bin-dir BIN_DIR]         Directory to place symlink (default: $BIN_DIR)"
   echo "  [--envs-dir ENVS_DIR]       Directory where environments are stored (default: $ENVS_DIR)"
   echo "  [--help]                    Show this usage message"
@@ -40,6 +46,14 @@ while [ $# -gt 0 ]; do
       ENTRYPOINT="$2"
       shift 2
       ;;
+    --name)
+      NAME="$2"
+      shift 2
+      ;;
+    --version)
+      VERSION="$2"
+      shift 2
+      ;;
     --bin-dir)
       BIN_DIR="$2"
       shift 2
@@ -61,7 +75,12 @@ done
 # -----------------------
 # 3. Rest of the script
 # -----------------------
-BIN="$BIN_DIR/$ENTRYPOINT"
+
+if [ -z "$NAME" ]; then
+  NAME="$ENTRYPOINT"
+fi
+
+BIN="$BIN_DIR/$NAME"
 ENV_DIR="$ENVS_DIR/$PROJECT"
 
 # check if the bin exists and error if it does
@@ -74,17 +93,23 @@ mkdir -p "$BIN_DIR"
 mkdir -p "$ENV_DIR"
 
 get_version_from_github() {
-  ver=$(wget -qO - "https://api.github.com/repos/${GH_USER}/${PROJECT}/releases/latest" |
-    grep '"tag_name":' | sed -E 's/.*"([^"]+)".*/\1/')
+  ver=$(wget -qO - "https://api.github.com/repos/${GH_USER}/${PROJECT}/releases/latest" \
+    | grep '"tag_name":' \
+    | sed -E 's/.*"([^"]+)".*/\1/')
   echo "$ver"
 }
 
-VERSION=$(get_version_from_github)
+# If user didn't provide --version, ask GitHub for the latest
+if [ -z "$VERSION" ]; then
+  VERSION=$(get_version_from_github)
+fi
+
 if [ -z "$VERSION" ]; then
   echo "Failed to get the latest version from GitHub."
   exit 1
 fi
-echo "Latest version: $VERSION"
+
+echo "Version: $VERSION"
 
 get_operating_system() {
   case "$(uname -s)" in
@@ -100,6 +125,7 @@ get_operating_system() {
       ;;
   esac
 }
+
 OS=$(get_operating_system)
 if [ -z "$OS" ]; then
   echo "Failed to determine the operating system."
@@ -121,6 +147,7 @@ get_architecture() {
       ;;
   esac
 }
+
 ARCH=$(get_architecture)
 if [ -z "$ARCH" ]; then
   echo "Failed to determine the architecture."
@@ -135,29 +162,37 @@ download_installer() {
 }
 
 URL="https://github.com/${GH_USER}/${PROJECT}/releases/download/${VERSION}/${PROJECT}-${VERSION}-${OS}-${ARCH}.sh"
+FILE="${ENV_DIR}/${PROJECT}-${VERSION}-${OS}-${ARCH}.sh"
 
-echo "Downloading installer from $URL"
-download_installer "${ENV_DIR}/${PROJECT}-${VERSION}-${OS}-${ARCH}.sh" "$URL"
+if [ -f "$FILE" ]; then
+  echo "Installer already downloaded: $FILE"
+else
+  echo "Installer not found, downloading..."
+  download_installer "$FILE" "$URL"
+fi
 
-# run the installer
+if [ ! -f "$FILE" ]; then
+  echo "Failed to download the installer."
+  exit 1
+fi
+
 echo "Running installer"
-chmod +x "${ENV_DIR}/${PROJECT}-${VERSION}-${OS}-${ARCH}.sh"
-"${ENV_DIR}/${PROJECT}-${VERSION}-${OS}-${ARCH}.sh" --output-directory "${ENV_DIR}"
+chmod +x "$FILE"
+"$FILE" --output-directory "$ENV_DIR"
 
 # add the entrypoint to activate.sh
-cat "${ENV_DIR}/activate.sh" > "${ENV_DIR}/${ENTRYPOINT}" 
-echo "$ENTRYPOINT \$@" >> "${ENV_DIR}/${ENTRYPOINT}"
-chmod +x "${ENV_DIR}/${ENTRYPOINT}"
+cat "${ENV_DIR}/activate.sh" > "${ENV_DIR}/${NAME}"
+echo "$ENTRYPOINT \$@" >> "${ENV_DIR}/${NAME}"
+chmod +x "${ENV_DIR}/${NAME}"
 
 echo "Creating symlink to $ENTRYPOINT in $BIN_DIR"
 ln -s "${ENV_DIR}/${ENTRYPOINT}" "$BIN"
 
 # check BIN_DIR is in PATH
 if ! echo "$PATH" | grep -q "$BIN_DIR"; then
-  echo "Error: $BIN_DIR is not in your PATH. Please add it to your PATH."
+  echo "Warning: $BIN_DIR is not in your PATH. Please add it to your PATH."
   echo "You can do this by adding the following line to your ~/.bashrc or ~/.zshrc:"
   echo "export PATH=\$PATH:$BIN_DIR"
-  exit 1
 fi
 
-echo "Installation complete. You can now run '$ENTRYPOINT' from anywhere."
+echo "Installation complete!"
